Switch to Slim 4, Illuminate container
Revised service configuration a bit to take advantage of Illuminate
Container autowiring, but was able to keep the services themselves, as
well as routes, as-is, with the exception of the cookie handler function
needing to be wrapped in a class.

As part of this jump, bumped Alpine up to the newest version so we're
now running PHP 7.3 by default.